refactor: rename suri_cmd.py
